fn_code/fn_proj_budget_deleter/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub
google-cloud-resource-manager==1.13.0
google-cloud-logging==3.11.3